Adds implementation files to remove linker errors
This results in typedef errors in libsnark files. The reason is unclear,
but it looks like these files are removed and recreated when the project
is being built. Either the files are not found since they aren't there
or perhaps not included correctly in some of the build definition files.

One source of error could be also that in default_types/r1cs_ppzksnark_pp.hpp
the file default_ec_pp.hpp is in the same directory, but it is referenced
with a different path. This alludes the functioning builds on other systems
somehow rebaase or otherwise alter the inclusion paths.